backend/requirements.txt
fastapi>=0.100.0
uvicorn==0.24.0
yfinance==0.2.28
sqlalchemy==2.0.23
pydantic==2.5.0
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
feedparser==6.0.10
beautifulsoup4==4.12.2
pandas>=2.2.0
numpy==1.26.2
python-dateutil==2.8.2
python-dotenv==1.0.0